Fixed editing of context menu items

+++ apps/web/src/routes/AuthLayoutRouter/components/ContextMenu/ContextMenu.jsx
@@ -66,7 +66,7 @@ export default function ContextMenu (props) {
       return getStaticMenuWidgets({ isPublicContext, isMyContext: isMyContext || isAllContext, profileUrl })
     }
     return getContextWidgets(state, group)
-  }, (a, b) => a.length === b.length && a.every((widget, index) => widget.id === b[index].id && widget.highlightNumber === b[index].highlightNumber))
+  })
 
   const contextWidgets = useMemo(() => {
     return rawContextWidgets.map(widget => ContextWidgetPresenter(widget))
@@ -79,7 +79,9 @@ export default function ContextMenu (props) {
     return false
   }, [group, isMyContext, isPublicContext, isAllContext])
 
-  const orderedWidgets = useMemo(() => orderContextWidgetsForContextMenu(contextWidgets), [contextWidgets])
+  const orderedWidgets = useMemo(() => {
+    return orderContextWidgetsForContextMenu(contextWidgets)
+  }, [contextWidgets])
 
   const isEditing = getQuerystringParam('cme', location) === 'yes' && canAdminister
 
@@ -100,6 +102,8 @@ export default function ContextMenu (props) {
   [currentWidgetIds])
   previousWidgetIds = currentWidgetIds
   const newWidgetRef = useRef()
+
+  const [forceUpdate, setForceUpdate] = useState({})
 
   useEffect(() => {
     if (isEditing) {
@@ -138,9 +142,12 @@ export default function ContextMenu (props) {
 
   const handleDragEnd = (event) => {
     const { active, over } = event
+
     setIsDragging(false)
+
     if (over && over.id !== active.id && over.id !== 'remove') {
       const orderInFrontOfWidget = over.data?.current?.widget
+
       dispatch(updateContextWidget({
         contextWidgetId: active.id,
         groupId: group.id,
@@ -151,12 +158,21 @@ export default function ContextMenu (props) {
           remove: over.id === 'remove'
         }
       }))
+
+      // Force a re-render with a new object reference
+      setForceUpdate({})
     }
-    if (over.id === 'remove') {
+
+    if (over && over.id === 'remove') {
       dispatch(removeWidgetFromMenu({ contextWidgetId: active.id, groupId: group.id }))
     }
+
     setActiveWidget(null)
   }
+
+  useEffect(() => {
+    // This empty dependency array ensures the effect runs when forceUpdate changes
+  }, [forceUpdate])
 
   return (
     <div className={cn('ContextMenu bg-background z-20 overflow-y-auto h-lvh w-[300px] shadow-md', { [classes.mapView]: mapView }, { [classes.showGroupMenu]: isNavOpen }, className)}>
@@ -190,6 +206,7 @@ export default function ContextMenu (props) {
           <DndContext onDragStart={handleDragStart} onDragEnd={handleDragEnd} collisionDetection={closestCorners} modifiers={[restrictToVerticalAxis]}>
             <div className='w-full'>
               <ContextWidgetList
+                key={JSON.stringify(forceUpdate)}
                 isDragging={isDragging}
                 isEditing={isEditing}
                 contextWidgets={orderedWidgets}
@@ -388,7 +405,7 @@ function ContextMenuItem ({ widget, groupSlug, rootPath, canAdminister = false, 
                   <SpecialTopElementRenderer widget={widget} group={group} isEditing={isEditing} />
                   <ul className='p-0'>
                     {loading && <li key='loading'>Loading...</li>}
-                    {presentedlistItems.length > 0 && presentedlistItems.map(item => <ListItemRenderer key={item.id} item={item} rootPath={rootPath} groupSlug={groupSlug} isDragging={isDragging} canDnd={canDnd} activeWidget={activeWidget} invalidChild={isInvalidChild} handlePositionedAdd={handlePositionedAdd} />)}
+                    {presentedlistItems.length > 0 && presentedlistItems.map(item => <ListItemRenderer key={item.id + '-' + Date.now()} item={item} rootPath={rootPath} groupSlug={groupSlug} isDragging={isDragging} canDnd={canDnd} activeWidget={activeWidget} invalidChild={isInvalidChild} handlePositionedAdd={handlePositionedAdd} />)}
                     {widget.id && isEditing && !['home', 'setup'].includes(widget.type) &&
                       <li>
                         <DropZone isDragging={isDragging} hide={hideDropZone || hideBottomDropZone} isDroppable={canDnd && !url} droppableParams={{ id: 'bottom-of-child-list' + widget.id, data: { addToEnd: true, parentId: widget.id } }}>
@@ -406,7 +423,7 @@ function ContextMenuItem ({ widget, groupSlug, rootPath, canAdminister = false, 
                   <SpecialTopElementRenderer widget={widget} group={group} isEditing={isEditing} />
                   <ul className='px-1 pt-1 pb-2'>
                     {loading && presentedlistItems.length === 0 && <li key='loading'>Loading...</li>}
-                    {presentedlistItems.length > 0 && presentedlistItems.map(item => <ListItemRenderer key={item.id} item={item} rootPath={rootPath} groupSlug={groupSlug} isDragging={isDragging} canDnd={canDnd} activeWidget={activeWidget} invalidChild={isInvalidChild} handlePositionedAdd={handlePositionedAdd} />)}
+                    {presentedlistItems.length > 0 && presentedlistItems.map(item => <ListItemRenderer key={item.id + '-' + Date.now()} item={item} rootPath={rootPath} groupSlug={groupSlug} isDragging={isDragging} canDnd={canDnd} activeWidget={activeWidget} invalidChild={isInvalidChild} handlePositionedAdd={handlePositionedAdd} />)}
                   </ul>
                 </div>}
             </div>)}
@@ -473,7 +490,7 @@ function ListItemRenderer ({ item, rootPath, groupSlug, canDnd, isOverlay = fals
 
   const isActive = item.viewUser?.lastActiveAt ? new Date(parseInt(item.viewUser.lastActiveAt)) > new Date(Date.now() - 1000 * 60 * 4) : false
   return (
-    <React.Fragment key={item.id + itemTitle}>
+    <React.Fragment key={item.id + "-" + Date.now()}>
       <DropZone hide={hideDropZone || invalidChild || !canDnd} droppableParams={{ id: `${item.id}`, data: { widget: item } }}>
         &nbsp;
       </DropZone>
